PS-3256 - Introduce Status field on program management

+++ src/tenant/tenant.controller.ts
@@ -1,11 +1,16 @@
-import { Body, Controller, Delete, Get, Param, ParseUUIDPipe, Patch, Post, Query, Req, Res, SerializeOptions, UploadedFiles, UseInterceptors, UsePipes, ValidationPipe } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, Get, Param, ParseUUIDPipe, Patch, Post, Query, Req, Res, SerializeOptions, UploadedFiles, UseInterceptors, UsePipes, ValidationPipe } from '@nestjs/common';
 import { TenantService } from './tenant.service';
-import { ApiCreatedResponse, ApiForbiddenResponse, ApiQuery } from '@nestjs/swagger';
+import { ApiCreatedResponse, ApiForbiddenResponse, ApiQuery, ApiTags } from '@nestjs/swagger';
 import { TenantCreateDto } from './dto/tenant-create.dto';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { FilesUploadService } from 'src/common/services/upload-file';
 import { TenantUpdateDto } from './dto/tenant-update.dto';
-import { Response } from "express";
+import { Request, Response } from "express";
+import { TenantSearchDTO } from './dto/tenant-search.dto';
+import { API_RESPONSES } from '@utils/response.messages';
+import { isUUID } from 'class-validator';
+
+@ApiTags("Tenant")
 @Controller('tenant')
 export class TenantController {
     constructor(
@@ -14,8 +19,8 @@ export class TenantController {
     ) { }
     //Get tenant information
     @Get("/read")
-    @ApiCreatedResponse({ description: "Tenant Data Fetch" })
-    @ApiForbiddenResponse({ description: "Forbidden" })
+    @ApiCreatedResponse({ description: API_RESPONSES.TENANT_SEARCH_SUCCESS })
+    @ApiForbiddenResponse({ description: API_RESPONSES.FORBIDDEN })
     @UsePipes(ValidationPipe)
     @SerializeOptions({
         strategy: "excludeAll",
@@ -23,24 +28,41 @@ export class TenantController {
     public async getTenants(
         @Req() request: Request,
         @Res() response: Response
-    ) {
+    ): Promise<Response>{
         return await this.tenantService.getTenants(request, response);
     }
 
+    //Search Tenanr deatils
+    @Post("/search")
+    @ApiCreatedResponse({ description: API_RESPONSES.TENANT_SEARCH_SUCCESS })
+    @ApiForbiddenResponse({ description: API_RESPONSES.FORBIDDEN })
+    @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
+    public async searchTenants(
+        @Body() tenantSearchDTO: TenantSearchDTO,
+        @Req() request: Request,
+        @Res() response: Response,
+    ): Promise<Response> {
+        return this.tenantService.searchTenants(request, tenantSearchDTO, response);
+    }
+
+
     //Create a new tenant
     @Post("/create")
-    @ApiCreatedResponse({ description: "Tenant Created Successfully" })
-    @ApiForbiddenResponse({ description: "Forbidden" })
+    @ApiCreatedResponse({ description: API_RESPONSES.TENANT_CREATE })
+    @ApiForbiddenResponse({ description: API_RESPONSES.FORBIDDEN })
     @UseInterceptors(FilesInterceptor('programImages', 10))
     @UsePipes(ValidationPipe)
-    @ApiQuery({ name: "userId", required: false })
     public async createTenants(
         @Req() request: Request,
         @Res() response: Response,
         @Body() tenantCreateDto: TenantCreateDto,
         @UploadedFiles() files: Express.Multer.File[],
-        @Query("userId") userId: string | null = null
-    ) {
+        @Query("userId") userId: string
+    ): Promise<Response> {
+        if (!userId || !isUUID(userId)) {
+            throw new BadRequestException(API_RESPONSES.UNAUTHORIZED);
+        }
+
         const uploadedFiles = [];
 
         // Loop through each file and upload it
@@ -62,14 +84,20 @@ export class TenantController {
     @ApiForbiddenResponse({ description: "Forbidden" })
     @UseInterceptors(FilesInterceptor('programImages', 10))
     @UsePipes(ValidationPipe)
-    @ApiQuery({ name: "userId", required: false })
     public async updateTenants(
         @Res() response: Response,
         @Param("id", ParseUUIDPipe) id: string,
         @Body() tenantUpdateDto: TenantUpdateDto,
         @UploadedFiles() files: Express.Multer.File[],
-        @Query("userId") userId: string | null = null,
-    ) {
+        @Query("userId") userId: string,
+    ): Promise<Response> {
+        const tenantId = id;        
+        if (!userId || !isUUID(userId)) {
+            throw new BadRequestException(API_RESPONSES.UNAUTHORIZED);
+        }
+        if (!tenantId || !isUUID(tenantId)) {
+            throw new BadRequestException(API_RESPONSES.REQUIRED_AND_UUID);
+        }
         const uploadedFiles = [];
 
         // Loop through each file and upload it
@@ -81,8 +109,7 @@ export class TenantController {
             // Assuming tenantCreateDto needs an array of file paths
             tenantUpdateDto.programImages = uploadedFiles.map(file => file.filePath); // Adjust field as needed
         }
-        const tenantId = id;        
-        tenantUpdateDto.updatedBy = userId || null;
+        tenantUpdateDto.updatedBy = userId;
         return await this.tenantService.updateTenants(tenantId, tenantUpdateDto, response);
     }
 
@@ -98,9 +125,16 @@ export class TenantController {
     public async deleteTenants(
         @Req() request: Request,
         @Res() response: Response,
-        @Query("id") id: string
+        @Query("id") id: string,
+        @Query("userId") userId: string,
     ) {
-        const tenantId = id;
+        const tenantId = id;        
+        if (!userId || !isUUID(userId)) {
+            throw new BadRequestException(API_RESPONSES.UNAUTHORIZED);
+        }
+        if (!tenantId || !isUUID(tenantId)) {
+            throw new BadRequestException(API_RESPONSES.REQUIRED_AND_UUID);
+        }
         return await this.tenantService.deleteTenants(request, tenantId, response);
     }
